clase6/RS_232/ejemplo4/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
docutils==0.20.1
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
httptools==0.6.0
idna==3.4
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-examples==2.2.1
Kivy-Garden==0.1.5
Pillow==10.0.1
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
pypiwin32==223
pyserial==3.5
python-dotenv==1.0.0
pywin32==306
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
